[feat] 나의 식물 통계 모달 기능 추가

+++ src/main/java/org/joonzis/myplant/dto/MyPlantStatisticsDTO.java
@@ -11,9 +11,9 @@ import lombok.NoArgsConstructor;
 @Data
 public class MyPlantStatisticsDTO {
 	private int myplant_id;
-	private double temperature;
-	private double humidity;
-	private double illumination;
-	private double soil_moisture;
+	private Integer temperature;
+	private Integer humidity;
+	private Integer illumination;
+	private Integer soil_moisture;
 	private Date sensing_time;
 }
